phase1_requirements.txt
# Phase 1 Training Requirements
# Exact versions from gpu_setup.sh
# For H100 with CUDA 12.4, PyTorch 2.6.0
# Last updated: 2025-11-16

# Core utilities
psutil==6.0.0
packaging
rich
tensorboard
tqdm
jsonlines
ninja

# Quantization & Optimization
bitsandbytes==0.48.1
torchao==0.13.0

# Transformers Ecosystem
transformers==4.43.3
tokenizers==0.19.1

# Training Frameworks
peft==0.11.1
accelerate==1.11.0
trl==0.9.6
datasets==4.3.0

# Unsloth dependencies
einops
sentencepiece

# HuggingFace Hub & Serialization
huggingface-hub
safetensors

# Scientific Computing
scipy==1.14.0
scikit-learn==1.4.2
